requirements.txt
grpcio==1.62.0
matplotlib==3.8.3
numpy==1.26.4
opcua==0.98.13
pandas==2.2.1
protobuf==4.25.3
scikit_learn==1.4.1.post1
scipy==1.12.0
tslearn==0.6.3
